requirements.txt
streamlit
openai
pdfplumber
tavily-python
python-dotenv
